backend/requirements.txt
# Flask Backend Requirements for FitSpaceAI
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0

# Optional dependencies for production
# Flask-SQLAlchemy==3.1.1  # For database integration
# Flask-JWT-Extended==4.5.3  # For JWT authentication
# Redis==5.0.1  # For caching
# Celery==5.3.4  # For background tasks
